+++ utils/create-gear-app/src/hooks/index.ts
@@ -1,3 +1,5 @@
+import { useAccounts, useLoggedInAccount } from './accounts';
+import { useBalanceSubscription } from './api';
 import { useAccount, useApi } from './context';
 
-export { useAccount, useApi };
+export { useAccounts, useLoggedInAccount, useBalanceSubscription, useAccount, useApi };
